Successfully snap to the grid

+++ Towers2020/ItemTile.cpp
@@ -35,8 +35,8 @@ void CItemTile::SetGridLocation(int x, int y)
 	mGridPositionX = x;
 	mGridPositionY = y;
 
-	double xPixelLocation = x * TileWidth;
-	double yPixelLocation = y * TileHeight;
+	double xPixelLocation = double(x) * TileWidth;
+	double yPixelLocation = double(y) * TileHeight;
 
 	SetLocation(xPixelLocation, yPixelLocation);
 }
